requirements.txt
Flask==3.0.1
gunicorn==21.2.0

requests==2.31.0
structlog==24.1.0

google-auth==2.3.2

fastapi==0.110.1
langchain==0.1.16
langchain_community==0.0.32
langchain_core==0.1.42
google-auth==2.27.0
langchain_google_genai==1.0.2
langserve==0.0.51
pydantic==2.7.0
python-dotenv==1.0.1
uvicorn==0.29.0
langchain-google-community==1.0.2
google-cloud-storage==2.14.0
unstructured==0.13.2
unstructured-client==0.18.0
unstructured[pdf]
